refactor: 페이지에서 피드에디터 분리
